Create ThreadLocal
This abstracts the ThreadAlloc concept into an parameterised ThreadLocal
class.  The class enables registering the teardown operation off the
fast path.
