+++ src/snmalloc/backend/fixedglobalconfig.h
@@ -69,15 +69,6 @@ namespace snmalloc
       return opts;
     }();
 
-    // This needs to be a forward reference as the
-    // thread local state will need to know about this.
-    // This may allocate, so must be called once a thread
-    // local allocator exists.
-    static void register_clean_up()
-    {
-      snmalloc::register_clean_up();
-    }
-
     static void init(LocalState* local_state, void* base, size_t length)
     {
       UNUSED(local_state);
